+++ semo-core/src/main/java/sandbox/semo/domain/member/dto/request/MemberFormDecision.java
@@ -1,4 +1,4 @@
-package sandbox.semo.domain.form.dto.request;
+package sandbox.semo.domain.member.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
