+++ semo-core/src/main/java/sandbox/semo/domain/member/dto/request/MemberFormRegister.java
@@ -1,4 +1,4 @@
-package sandbox.semo.domain.form.dto.request;
+package sandbox.semo.domain.member.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
