+++ semo-core/src/main/java/sandbox/semo/domain/member/repository/MemberFormRepository.java
@@ -1,7 +1,8 @@
-package sandbox.semo.domain.form.repository;
+package sandbox.semo.domain.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sandbox.semo.domain.form.entity.MemberForm;
+import sandbox.semo.domain.member.entity.MemberForm;
 
 public interface MemberFormRepository extends JpaRepository<MemberForm, Long> {
+
 }
